[FEAT]: read from local storage lo load books
